fix: improve overall time execution
Optimize the time execution focusing on two different aspects
regarding the way we match languages with a file:

 - Remove the use of the 'globSync' function. This function is
called each time a wildcard pattern is found when adding a new
language. The function analyzes every time all the files inside
the directory, and gets really slow when the number of files
increase substantially. In the second bullet point it is explained
how it can be fully removed.

 - There is redundant work being done twice: matching languages
with a single file. First, it is done in the 'globSync' when
adding a new language (matched against all files). Secondly, it
is done in the 'getLanguageMatches'. Therefore, as all public
methods call the 'getLanguageMatches' function, we just need to
check in it which languages match with the current file, without
the need to precompute it previously in the 'replaceGlobPatternsInFilenames'
function.